pyproject.toml
[project]
name = "mad-td"
version = "0.1.0"
description = "MAD-TD codebase"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "brax>=0.12.4",
    "distrax>=0.1.5",
    "dm-control>=1.0.28",
    "flax>=0.10.4",
    "gymnasium>=1.1.1",
    "gymnax>=0.0.8",
    "hydra-core>=1.3.2",
    "optax>=0.2.4",
    "tqdm>=4.67.1",
    "wandb>=0.19.8",
    "imageio==2.14.1",
]
